requirements.txt
matplotlib==3.10.9
numpy==2.4.4
onnx==1.21.0
onnxruntime==1.25.1
Pillow==12.2.0
scikit_learn==1.8.0
torch==2.11.0
torchvision==0.26.0

opencv-python